sakha_backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
firebase-admin==6.0.1
langchain==0.0.142
openai==0.27.0
pinecone-client==2.2.3
redis==4.5.5
django-redis==5.2.0
pymongo==4.3.3
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
